// src/generated/t_tech/invest/grpc/users.ts
export interface GetUserTariffResponse {
  /** Массив лимитов пользователя по unary-запросам. */
  unaryLimits: UnaryLimit[];
  /** Массив лимитов пользователей для stream-соединений. */
  streamLimits: StreamLimit[];
}

/** Лимит unary-методов. */
export interface UnaryLimit {
  /** Количество unary-запросов в минуту. */
  limitPerMinute: number;
  /** Названия методов. */
  methods: string[];
  /** Количество unary-запросов в секунду. */
  limitPerSecond?: number | undefined;
}

/** Лимит stream-соединений. */
export interface StreamLimit {
  /** Максимальное количество stream-соединений. */
  limit: number;
  /** Названия stream-методов. */
  streams: string[];
  /** Текущее количество открытых stream-соединений. */
  open: number;
}
// ...
export const GetUserTariffResponse: MessageFns<GetUserTariffResponse> = {
  encode(message: GetUserTariffResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.unaryLimits) {
      UnaryLimit.encode(v!, writer.uint32(10).fork()).join();
    }
    for (const v of message.streamLimits) {
      StreamLimit.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetUserTariffResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetUserTariffResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.unaryLimits.push(UnaryLimit.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.streamLimits.push(StreamLimit.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetUserTariffResponse {
    return {
      unaryLimits: globalThis.Array.isArray(object?.unaryLimits)
        ? object.unaryLimits.map((e: any) => UnaryLimit.fromJSON(e))
        : globalThis.Array.isArray(object?.unary_limits)
        ? object.unary_limits.map((e: any) => UnaryLimit.fromJSON(e))
        : [],
      streamLimits: globalThis.Array.isArray(object?.streamLimits)
        ? object.streamLimits.map((e: any) => StreamLimit.fromJSON(e))
        : globalThis.Array.isArray(object?.stream_limits)
        ? object.stream_limits.map((e: any) => StreamLimit.fromJSON(e))
        : [],
    };
  },

  toJSON(message: GetUserTariffResponse): unknown {
    const obj: any = {};
    if (message.unaryLimits?.length) {
      obj.unaryLimits = message.unaryLimits.map((e) => UnaryLimit.toJSON(e));
    }
    if (message.streamLimits?.length) {
      obj.streamLimits = message.streamLimits.map((e) => StreamLimit.toJSON(e));
    }
    return obj;
  },

  create(base?: DeepPartial<GetUserTariffResponse>): GetUserTariffResponse {
    return GetUserTariffResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetUserTariffResponse>): GetUserTariffResponse {
    const message = createBaseGetUserTariffResponse();
    message.unaryLimits = object.unaryLimits?.map((e) => UnaryLimit.fromPartial(e)) || [];
    message.streamLimits = object.streamLimits?.map((e) => StreamLimit.fromPartial(e)) || [];
    return message;
  },
};

function createBaseUnaryLimit(): UnaryLimit {
  return { limitPerMinute: 0, methods: [], limitPerSecond: undefined };
}

export const UnaryLimit: MessageFns<UnaryLimit> = {
  encode(message: UnaryLimit, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.limitPerMinute !== 0) {
      writer.uint32(8).int32(message.limitPerMinute);
    }
    for (const v of message.methods) {
      writer.uint32(18).string(v!);
    }
    if (message.limitPerSecond !== undefined) {
      writer.uint32(24).int32(message.limitPerSecond);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): UnaryLimit {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUnaryLimit();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.limitPerMinute = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.methods.push(reader.string());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.limitPerSecond = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UnaryLimit {
    return {
      limitPerMinute: isSet(object.limitPerMinute)
        ? globalThis.Number(object.limitPerMinute)
        : isSet(object.limit_per_minute)
        ? globalThis.Number(object.limit_per_minute)
        : 0,
      methods: globalThis.Array.isArray(object?.methods) ? object.methods.map((e: any) => globalThis.String(e)) : [],
      limitPerSecond: isSet(object.limitPerSecond)
        ? globalThis.Number(object.limitPerSecond)
        : isSet(object.limit_per_second)
        ? globalThis.Number(object.limit_per_second)
        : undefined,
    };
  },

  toJSON(message: UnaryLimit): unknown {
    const obj: any = {};
    if (message.limitPerMinute !== 0) {
      obj.limitPerMinute = Math.round(message.limitPerMinute);
    }
    if (message.methods?.length) {
      obj.methods = message.methods;
    }
    if (message.limitPerSecond !== undefined) {
      obj.limitPerSecond = Math.round(message.limitPerSecond);
    }
    return obj;
  },

  create(base?: DeepPartial<UnaryLimit>): UnaryLimit {
    return UnaryLimit.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<UnaryLimit>): UnaryLimit {
    const message = createBaseUnaryLimit();
    message.limitPerMinute = object.limitPerMinute ?? 0;
    message.methods = object.methods?.map((e) => e) || [];
    message.limitPerSecond = object.limitPerSecond ?? undefined;
    return message;
  },
};

function createBaseStreamLimit(): StreamLimit {
  return { limit: 0, streams: [], open: 0 };
}

export const StreamLimit: MessageFns<StreamLimit> = {
  encode(message: StreamLimit, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.limit !== 0) {
      writer.uint32(8).int32(message.limit);
    }
    for (const v of message.streams) {
      writer.uint32(18).string(v!);
    }
    if (message.open !== 0) {
      writer.uint32(24).int32(message.open);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): StreamLimit {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseStreamLimit();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.limit = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.streams.push(reader.string());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.open = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): StreamLimit {
    return {
      limit: isSet(object.limit) ? globalThis.Number(object.limit) : 0,
      streams: globalThis.Array.isArray(object?.streams) ? object.streams.map((e: any) => globalThis.String(e)) : [],
      open: isSet(object.open) ? globalThis.Number(object.open) : 0,
    };
  },

  toJSON(message: StreamLimit): unknown {
    const obj: any = {};
    if (message.limit !== 0) {
      obj.limit = Math.round(message.limit);
    }
    if (message.streams?.length) {
      obj.streams = message.streams;
    }
    if (message.open !== 0) {
      obj.open = Math.round(message.open);
    }
    return obj;
  },

  create(base?: DeepPartial<StreamLimit>): StreamLimit {
    return StreamLimit.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<StreamLimit>): StreamLimit {
    const message = createBaseStreamLimit();
    message.limit = object.limit ?? 0;
    message.streams = object.streams?.map((e) => e) || [];
    message.open = object.open ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}